move patch rich relationships into single test file
